Created an array for tasks and called the add
Great work creating objects and multiple functions for adding/deleting tasks. One thing I added was creating an array of objects instead of defining each property as its own variable. This will make it easier to work with as you build your application. I put a little confirmation that the user has added a task as well which is a nice thing for a user to see.